Backport to Qt5

+++ lib/callofsuli-core/utils_.h
@@ -51,8 +51,8 @@
 #define JSON_TO_INTEGER(x)	x.toInteger()
 #define JSON_TO_INTEGER_Y(x, y)	x.toInteger(y)
 #else
-#define JSON_TO_INTEGER(x)	x.toInt()
-#define JSON_TO_INTEGER_Y(x, y)	x.toInt(y)
+#define JSON_TO_INTEGER(x)	x.toDouble()
+#define JSON_TO_INTEGER_Y(x, y)	x.toDouble(y)
 #endif
